HydrOGEnMod

Documentation for HydrOGEnMod.

HydrOGEnMod.build_complementarity_model

HydrOGEnMod.build_optimization_model

HydrOGEnMod.get_HydrOGEnMod_data

HydrOGEnMod.plot_model_nodes

HydrOGEnMod.plot_model_nodes

HydrOGEnMod.plot_model_nodes

HydrOGEnMod.plot_model_nodes

HydrOGEnMod.plot_model_nodes

HydrOGEnMod.solve_mcp

HydrOGEnMod.write_csv

HydrOGEnMod.write_csv

JuMP.optimize!

HydrOGEnMod.build_complementarity_model — Function

build_complementarity_model(data::ModelData, BIG = 1e+6, logfile = "")

Build a JuMP model with complementarity extension from the underlying model data.

A logfile and upper bounds on variables can be specified.

Examples

julia> using HydrOGEnMod
julia> data = get_HydrOGEnMod_data();
julia> build_complementarity_model(data)
A JuMP Model
Feasibility problem with:
Variables: 35
`JuMP.VariableRef`-in-`MathOptInterface.GreaterThan{Float64}`: 34 constraints
`JuMP.VariableRef`-in-`MathOptInterface.LessThan{Float64}`: 34 constraints
Model mode: AUTOMATIC
CachingOptimizer state: NO_OPTIMIZER
Solver name: No optimizer attached.
Names registered in the model: arc_capacity, conversion_capacity, conversion_cost, conversion_expansion_cost, conversion_repurposing_cost, demand, equal_arc_expansion, equal_arc_repurposing, gross_surplus, input_block_limit, input_cost, input_expansion_cost, input_expansion_limit, market_power_adjustment, prices, producer_mass_balance, production_capacity, production_cost, production_expansion_cost, production_expansion_limit, q_I, q_P_T, q_S, q_S_in, q_S_out, q_T, q_T_D, q_V, storage_capacities, storage_cost, storage_expansion_cost, storage_expansion_limit, storage_mass_balance, storage_repurposing_cost, trade_sanctions, trader_mass_balance, transport_cost, transport_expansion_cost, transport_repurposing_cost, yearly_arc_flows, yearly_demand, yearly_input_procurement, yearly_production, Δ_A, Δ_I, Δ_P, Δ_RA, Δ_RS, Δ_RV, Δ_S, Δ_V

HydrOGEnMod.build_optimization_model — Function

build_optimization_model(data::ModelData, BIG = 1e+6, logfile = "")

Build a JuMP optimization model (QP) from the underlying model data.

A logfile and upper bounds on variables can be specified.

Examples

julia-repl julia> using HydrOGEnMod julia> data = get_HydrOGEnMod_data(); julia> build_optimization_model(data) A JuMP Model Maximization problem with: Variables: 11 Objective function type: JuMP.QuadExpr `JuMP.AffExpr`-in-`MathOptInterface.EqualTo{Float64}`: 3 constraints `JuMP.AffExpr`-in-`MathOptInterface.LessThan{Float64}`: 8 constraints `JuMP.VariableRef`-in-`MathOptInterface.GreaterThan{Float64}`: 11 constraints `JuMP.VariableRef`-in-`MathOptInterface.LessThan{Float64}`: 11 constraints Model mode: AUTOMATIC CachingOptimizer state: NO_OPTIMIZER Solver name: No optimizer attached. Names registered in the model: arc_capacity, conversion_capacity, conversion_cost, conversion_expansion_cost, conversion_repurposing_cost, demand, equal_arc_expansion, equal_arc_repurposing, gross_surplus, input_block_limit, input_cost, input_expansion_cost, input_expansion_limit, market_power_adjustment, prices, producer_mass_balance, production_capacity, production_cost, production_expansion_cost, production_expansion_limit, q_I, q_P_T, q_S, q_S_in, q_S_out, q_T, q_T_D, q_V, storage_capacities, storage_cost, storage_expansion_cost, storage_expansion_limit, storage_mass_balance, storage_repurposing_cost, trade_sanctions, trader_mass_balance, transport_cost, transport_expansion_cost, transport_repurposing_cost, yearly_arc_flows, yearly_demand, yearly_input_procurement, yearly_production, Δ_A, Δ_I, Δ_P, Δ_RA, Δ_RS, Δ_RV, Δ_S, Δ_V

HydrOGEnMod.get_HydrOGEnMod_data — Method

get_HydrOGEnMod_data(;data = joinpath(dirname(dirname(pathof(HydrOGEnMod))), "data", "data_pc"), logfile = "")

Get the model data

A separate datafile and a file where to write the log records can be specified via keyword arguments.

Examples

julia> using HydrOGEnMod
julia> data = get_HydrOGEnMod_data();

HydrOGEnMod.plot_model_nodes — Method

plot_model_nodes(arcs::Vector{TransportArc}; nodes_path = joinpath(@__DIR__, "maps", "WB_countries_Admin0_modified.geojson"), disputed_path = joinpath(@__DIR__, "maps", "WB_Admin0_disputed_areas_modified.geojson"), centroids_path = joinpath(@__DIR__, "maps", "centroids.csv"), colorscheme = :seaborn_colorblind6, unassigned_color = :gray85, strokecolor = :black, strokewidth = 0.1, linewidth = 0.2, linecolor = :black, size = (1200, 800), backgroundcolor = RGBf(0.95, 0.95, 0.95), rasterize = 10, map_title = "Model Nodes", save_path = "")

Plot the passed arcs into the model node map.

Separate method implemented for also passing arc values.

Keyword Arguments

nodes_path = joinpath(@__DIR__, "maps", "WB_countries_Admin0_modified.geojson"): Path where to find model nodes geojson file.

disputed_path = joinpath(@__DIR__, "maps", "WB_Admin0_disputed_areas_modified.geojson"): Path where to find disputed areas geojson file.

centroids_path = joinpath(@__DIR__, "maps", "centroids.csv"): Path where to find model region centroids file.

colorscheme = :seaborn_colorblind6: Colorscheme for the model regions.

unassigned_color = :gray85: Colorscheme for unassigned and disputed regions.

strokecolor = :black: Strokecolor around regions.

strokewidth = 0.1: Strokewidth around regions.

linewidth = 0.2: Arc linewidth connecting nodes.

linecolor = :black: Color of arcs connecting nodes.

size = (1200, 800): Picture size.

backgroundcolor = RGBf(0.95, 0.95, 0.95): Backgroundcolor.

rasterize = 10: Rasterization to reduce output file size, lesser implies smaller files.

map_title = "Model Nodes": Specification of the title.

save_path = "": Where to save the resulting graphic. Ending of filename automatically implies file format.

Examples

HydrOGEnMod.plot_model_nodes — Method

plot_model_nodes(; nodes_path = joinpath(@__DIR__, "maps", "WB_countries_Admin0_modified.geojson"), disputed_path = joinpath(@__DIR__, "maps", "WB_Admin0_disputed_areas_modified.geojson"), colorscheme = :seaborn_colorblind6, unassigned_color = :gray85, strokecolor = :black, strokewidth = 0.1, size = (1200, 800), backgroundcolor = RGBf(0.95, 0.95, 0.95), rasterize = 10, map_title = "Model Nodes", legend_title = "Node Names", save_path = "",)

Plot an overview of the model nodes.

Only active, if no positional argument is passed.

Keyword Arguments

nodes_path = joinpath(@__DIR__, "maps", "WB_countries_Admin0_modified.geojson"): Path where to find model nodes geojson file.

disputed_path = joinpath(@__DIR__, "maps", "WB_Admin0_disputed_areas_modified.geojson"): Path where to find disputed areas geojson file.

colorscheme = :seaborn_colorblind6: Colorscheme for the model regions.

unassigned_color = :gray85: Colorscheme for unassigned and disputed regions.

strokecolor = :black: Strokecolor around regions.

strokewidth = 0.1: Strokewidth around regions.

size = (1200, 800): Picture size.

backgroundcolor = RGBf(0.95, 0.95, 0.95): Backgroundcolor.

rasterize = 10: Rasterization to reduce output file size, lesser implies smaller files.

map_title = "Model Nodes": Specification of the title.

legend_title = "Node Names": Title of the legend.

save_path = "": Where to save the resulting graphic. Ending of filename automatically implies file format.

Examples

HydrOGEnMod.plot_model_nodes — Method

plot_model_nodes(container::JuMP.Containers.DenseAxisArray{T,N,Ax,L}; nodes_path = joinpath(@__DIR__, "maps", "WB_countries_Admin0_modified.geojson"), disputed_path = joinpath(@__DIR__, "maps", "WB_Admin0_disputed_areas_modified.geojson"), colorscheme = :plasma, unassigned_color = :gray85, strokecolor = :black, strokewidth = 0.1, size = (1200, 800), backgroundcolor = RGBf(0.95, 0.95, 0.95), rasterize = 10, map_title = "Model Nodes", save_path = "") where {T,N,Ax,L<:NTuple{N,JuMP.Containers._AxisLookup}}

Plot the values inside container into the model node map.

Separate methods implemented for DenseAxisArray and SparseAxisArray.

Keyword Arguments

nodes_path = joinpath(@__DIR__, "maps", "WB_countries_Admin0_modified.geojson"): Path where to find model nodes geojson file.

disputed_path = joinpath(@__DIR__, "maps", "WB_Admin0_disputed_areas_modified.geojson"): Path where to find disputed areas geojson file.

colorscheme = :plasma: Colorscheme for the model regions.

unassigned_color = :gray85: Colorscheme for unassigned and disputed regions.

strokecolor = :black: Strokecolor around regions.

strokewidth = 0.1: Strokewidth around regions.

size = (1200, 800): Picture size.

backgroundcolor = RGBf(0.95, 0.95, 0.95): Backgroundcolor.

rasterize = 10: Rasterization to reduce output file size, lesser implies smaller files.

map_title = "Model Nodes": Specification of the title.

save_path = "": Where to save the resulting graphic. Ending of filename automatically implies file format.

Examples

HydrOGEnMod.plot_model_nodes — Method

plot_model_nodes(container::JuMP.Containers.SparseAxisArray{T,N,K}; nodes_path = joinpath(@__DIR__, "maps", "WB_countries_Admin0_modified.geojson"), disputed_path = joinpath(@__DIR__, "maps", "WB_Admin0_disputed_areas_modified.geojson"), colorscheme = :plasma, unassigned_color = :gray85, strokecolor = :black, strokewidth = 0.1, size = (1200, 800), backgroundcolor = RGBf(0.95, 0.95, 0.95), rasterize = 10, map_title = "Model Nodes", save_path = "")

Plot the values inside container into the model node map.

Separate methods implemented for DenseAxisArray and SparseAxisArray.

Keyword Arguments

nodes_path = joinpath(@__DIR__, "maps", "WB_countries_Admin0_modified.geojson"): Path where to find model nodes geojson file.

disputed_path = joinpath(@__DIR__, "maps", "WB_Admin0_disputed_areas_modified.geojson"): Path where to find disputed areas geojson file.

colorscheme = :plasma: Colorscheme for the model regions.

unassigned_color = :gray85: Colorscheme for unassigned and disputed regions.

strokecolor = :black: Strokecolor around regions.

strokewidth = 0.1: Strokewidth around regions.

size = (1200, 800): Picture size.

backgroundcolor = RGBf(0.95, 0.95, 0.95): Backgroundcolor.

rasterize = 10: Rasterization to reduce output file size, lesser implies smaller files.

map_title = "Model Nodes": Specification of the title.

save_path = "": Where to save the resulting graphic. Ending of filename automatically implies file format.

Examples

HydrOGEnMod.plot_model_nodes — Method

plot_model_nodes(container::JuMP.Containers.DenseAxisArray{T,N,Ax,L}, arcs::Vector{TransportArc}; nodes_path = joinpath(@__DIR__, "maps", "WB_countries_Admin0_modified.geojson"), disputed_path = joinpath(@__DIR__, "maps", "WB_Admin0_disputed_areas_modified.geojson"), centroids_path = joinpath(@__DIR__, "maps", "centroids.csv"), colorscheme = :seaborn_colorblind6, unassigned_color = :gray85, strokecolor = :black, strokewidth = 0.1, linewidth = 1, linecolorscheme = :plasma, size = (1200, 800), backgroundcolor = RGBf(0.95, 0.95, 0.95), rasterize = 10, largestvalues = 10, map_title = "Model Nodes", save_path = "") where {T,N,Ax,L<:NTuple{N,JuMP.Containers._AxisLookup}}

Plot the passed arcs with coloring corresponding to given values in the container into the model node map.

Separate method implemented for just plotting normal values.

Keyword Arguments

nodes_path = joinpath(@__DIR__, "maps", "WB_countries_Admin0_modified.geojson"): Path where to find model nodes geojson file.

disputed_path = joinpath(@__DIR__, "maps", "WB_Admin0_disputed_areas_modified.geojson"): Path where to find disputed areas geojson file.

centroids_path = joinpath(@__DIR__, "maps", "centroids.csv"): Path where to find model region centroids file.

colorscheme = :seaborn_colorblind6: Colorscheme for the model regions.

unassigned_color = :gray85: Colorscheme for unassigned and disputed regions.

strokecolor = :black: Strokecolor around regions.

strokewidth = 0.1: Strokewidth around regions.

linewidth = 1: Arc linewidth connecting nodes.

linecolorscheme = :plasma: Colorscheme for coloring arcs connecting regions based on container values.

size = (1200, 800): Picture size.

backgroundcolor = RGBf(0.95, 0.95, 0.95): Backgroundcolor.

rasterize = 10: Rasterization to reduce output file size, lesser implies smaller files.

largestvalues = 10: The n largest arc values and equal to select.

minval = 1: The smallest value to include in the arc plot.

map_title = "Model Nodes": Specification of the title.

save_path = "": Where to save the resulting graphic. Ending of filename automatically implies file format.

Examples

HydrOGEnMod.solve_mcp — Method

solve_mcp(model::Model; logfile = "", kwargs...)

Solve a JuMP model with complementarity extension using the PATH solver.

A logfile can be specified, and solver specific options can be passed.

Examples

julia> using HydrOGEnMod
julia> data = get_HydrOGEnMod_data();
julia> model = build_complementarity_model(data);
julia> solve_mcp(model; output = "no", time_limit = 5)
Reading options file /var/folders/c7/86nwrvmd1c325s4lzwg77jz40000gn/T/jl_CJTzvb
Read of options file complete.

Path 5.1.99 (Mon Jul 24 14:47:28 2023)
Written by Todd Munson, Steven Dirkse, Youngdae Kim, and Michael Ferris
:Solved

HydrOGEnMod.write_csv — Method

write_csv(datadir, data::ModelData; logfile = "")

Write model data to a specific directory.

A logfile can be specified.

Examples

julia> using HydrOGEnMod
julia> data = get_HydrOGEnMod_data();
julia> write_csv("path/to/my/dir",data);

HydrOGEnMod.write_csv — Method

write_csv(resultsfile, model::Model; logfile = "")

Write model results to a specific directory.

A logfile can be specified.

Examples

julia> data = get_HydrOGEnMod_data();
julia> write_csv("path/to/my/dir",data);
julia> model = build_optimization_model(data);
julia> optimize!(model, Ipopt.Optimizer);
julia> write_csv("path/to/my/dir",model);

JuMP.optimize! — Method

optimize!(model::Model, optimizer; options..., logfile = "")

Solve a JuMP Model with the given optimizer.

A logfile can be specified, and solver specific options can be passed.

Examples

```julia-repl julia> using HydrOGEnMod julia> using Ipopt julia> data = getHydrOGEnModdata(); julia> model = buildoptimizationmodel(data); julia> optimize!(model, Ipopt.Optimizer, "maxwalltime" => 1e-5) ****************************************************************************** This program contains Ipopt, a library for large-scale nonlinear optimization. Ipopt is released as open source code under the Eclipse Public License (EPL). For more information visit https://github.com/coin-or/Ipopt ******************************************************************************

This is Ipopt version 3.14.13, running with linear solver MUMPS 5.6.0.

Number of nonzeros in equality constraint Jacobian...: 6 Number of nonzeros in inequality constraint Jacobian.: 7 Number of nonzeros in Lagrangian Hessian.............: 2

Total number of variables............................: 11 variables with only lower bounds: 0 variables with lower and upper bounds: 11 variables with only upper bounds: 0 Total number of equality constraints.................: 3 Total number of inequality constraints...............: 8 inequality constraints with only lower bounds: 0 inequality constraints with lower and upper bounds: 0 inequality constraints with only upper bounds: 8

iter objective infpr infdu lg(mu) ||d|| lg(rg) alphadu alphapr ls 0 -9.0099910e-02 1.00e-02 1.00e+01 -1.0 0.00e+00 - 0.00e+00 0.00e+00 0

Number of Iterations....: 0

                               (scaled)                 (unscaled)

Objective...............: 9.0099909800000114e-02 -9.0099909800000114e-02 Dual infeasibility......: 1.0000000000000000e+01 1.0000000000000000e+01 Constraint violation....: 9.9999800000000003e-03 9.9999800000000003e-03 Variable bound violation: 0.0000000000000000e+00 0.0000000000000000e+00 Complementarity.........: 9.9999999010000995e+05 9.9999999010000995e+05 Overall NLP error.......: 9.9999999010000995e+05 9.9999999010000995e+05

Number of objective function evaluations = 1 Number of objective gradient evaluations = 1 Number of equality constraint evaluations = 1 Number of inequality constraint evaluations = 1 Number of equality constraint Jacobian evaluations = 1 Number of inequality constraint Jacobian evaluations = 1 Number of Lagrangian Hessian evaluations = 0 Total seconds in IPOPT = 0.000

EXIT: Maximum wallclock time exceeded. ┌ Status: Optimizing Model │ Progress: Solution Status TIMELIMIT. │ Time elapsed: 0.14s └ @ HydrOGEnMod /Users/lb/Desktop/git/HydrOGEnMod/src/optimizationmodel.jl:999 ┌ Status: Optimizing Model │ Progress: Returning Model. │ Time elapsed: 0.14s └ @ HydrOGEnMod /Users/lb/Desktop/git/HydrOGEnMod/src/optimizationmodel.jl:1002 A JuMP Model Maximization problem with: Variables: 11 Objective function type: JuMP.QuadExpr JuMP.AffExpr-in-MathOptInterface.EqualTo{Float64}: 3 constraints JuMP.AffExpr-in-MathOptInterface.LessThan{Float64}: 8 constraints JuMP.VariableRef-in-MathOptInterface.GreaterThan{Float64}: 11 constraints JuMP.VariableRef-in-MathOptInterface.LessThan{Float64}: 11 constraints Model mode: AUTOMATIC CachingOptimizer state: ATTACHEDOPTIMIZER Solver name: Ipopt Names registered in the model: arccapacity, conversioncapacity, conversioncost, conversionexpansioncost, conversionrepurposingcost, demand, equalarcexpansion, equalarcrepurposing, grosssurplus, inputblocklimit, inputcost, inputexpansioncost, inputexpansionlimit, marketpoweradjustment, prices, producermassbalance, productioncapacity, productioncost, productionexpansioncost, productionexpansionlimit, qI, qPT, qS, qSin, qSout, qT, qTD, qV, storagecapacities, storagecost, storageexpansioncost, storageexpansionlimit, storagemassbalance, storagerepurposingcost, tradesanctions, tradermassbalance, transportcost, transportexpansioncost, transportrepurposingcost, yearlyarcflows, yearlydemand, yearlyinputprocurement, yearlyproduction, ΔA, ΔI, ΔP, ΔRA, ΔRS, ΔRV, ΔS, Δ_V```